Test for uniqueness of settings files before upload
Minor change in logic in that the uniqueness is tested before we start
to upload any settings file to S3.
The change of "--deploy" to "--deploy-config" is minor and can be ignored
(as in users can use "--deploy" or just "--dep" or ...)
Logic is a bit simplified around default file handling.