Make an honest attempt to copy credentials
While read access to credentials is limited from the developer's command line,
a "user" running in an EC2 instance, for example during a run of a data
pipeline, has no such limiations. For the "publish" feature of our ETL,
we should try to copy credentials, from, say "production" to "production/current"
so that we do not have to populate the "*/current/config" folders with credentials.